requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
chromadb==0.4.22
sentence-transformers==2.3.1
PyPDF2==3.0.1
google-generativeai==0.3.2
groq==0.4.1
pydantic==2.5.3
pydantic-settings==2.1.0
